+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This release fixes some bugs and adds minor features.
 
 - Feat: Show GPIO names
 - Feat: Support more HTTP method for HTTP action
+- Upd: Add Version to REST-API
 - Fix: Memory leak in list GPIOs function for REST-API
 
 ***